+++ messages_test.go
@@ -62,82 +62,6 @@ var (
 	}
 )
 
-func Test_sortMessages(t *testing.T) {
-	type args struct {
-		msgs []types.Message
-	}
-	tests := []struct {
-		name     string
-		args     args
-		wantMsgs []types.Message
-	}{
-		{
-			"empty",
-			args{[]types.Message{}},
-			[]types.Message{},
-		},
-		{
-			"sort ok",
-			args{[]types.Message{
-				{Message: slack.Message{Msg: slack.Msg{
-					Timestamp: "1643425514",
-				}}},
-				{Message: slack.Message{Msg: slack.Msg{
-					Timestamp: "1643425511",
-				}}},
-			}},
-			[]types.Message{
-				{Message: slack.Message{Msg: slack.Msg{
-					Timestamp: "1643425511",
-				}}},
-				{Message: slack.Message{Msg: slack.Msg{
-					Timestamp: "1643425514",
-				}}},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			sortMessages(tt.args.msgs)
-			assert.Equal(t, tt.wantMsgs, tt.args.msgs)
-		})
-	}
-}
-
-func TestSlackDumper_convertMsgs(t *testing.T) {
-
-	type args struct {
-		sm []slack.Message
-	}
-	tests := []struct {
-		name string
-		args args
-		want []types.Message
-	}{
-		{
-			"ok",
-			args{[]slack.Message{
-				testMsg1.Message,
-				testMsg2.Message,
-				testMsg3.Message,
-			}},
-			[]types.Message{
-				testMsg1,
-				testMsg2,
-				testMsg3,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			sd := &SlackDumper{}
-			if got := sd.convertMsgs(tt.args.sm); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("SlackDumper.convertMsgs() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
 func TestSlackDumper_DumpMessages(t *testing.T) {
 	type fields struct {
 		Users     Users
