+++ types/message.go
@@ -1,6 +1,7 @@
 package types
 
 import (
+	"sort"
 	"time"
 
 	"github.com/rusq/slackdump/v2/internal/structures"
@@ -36,4 +37,19 @@ func (m Message) IsThreadParent() bool {
 // conversation.
 func (m Message) IsThreadChild() bool {
 	return m.IsThread() && m.Msg.ReplyCount == 0
+}
+
+func SortMessages(msgs []Message) {
+	sort.Slice(msgs, func(i, j int) bool {
+		return msgs[i].Timestamp < msgs[j].Timestamp
+	})
+}
+
+// ConvertMsgs converts a slice of slack.Message to []types.Message.
+func ConvertMsgs(sm []slack.Message) []Message {
+	msgs := make([]Message, len(sm))
+	for i := range sm {
+		msgs[i].Message = sm[i]
+	}
+	return msgs
 }
